Initial commit.
Added files from working on this project on the plane. Added reverse
polish calculator (from a previous project). Brute forcing every
possible ordering of numbers and every operation is done in the main
class for now. Calculator needs to be changed to make it suit the game
environment better, most importantly including brackets.

Code does not find solutions that require brackets in the game, and
reverse polish does not understand the 1 bracket limitation of the game.
Need to create junit tests to make development more straightforward.
Need to add userinputs and give option to filter out solutions that use
certain operators (as some levels in the game disable the use of some
operators).